Centralize Realm init
Alternate approach to managing Realm that removes the need for an additional
helper and provides a single Realm instance to the entire app. Hopefully
consolidates some code, but I still don't like the fact that models need to
receive a realm instance to talk to the db, or that providers and components
need to know about Realm at all.