Add a superclass for models that holds a reference to Realm
Very messy proof of concept that shows how we could load realm in a provider
and use a superclass of all models to hold a reference to realm so models can
be responsible for Realm CRUD instead of components. Also starts
encapsulating basic CRUD in ActiveRecord like methods (new, save, etc) so
models handle these operations and not helpers or providers.