Store current user logged in state in the User model
Removes MMKV as a dependency. This approach is slightly dangerous in that
syncing with the API might alter the signed in state. If that becomes a
problem, we might move to store this data in a separate model, or in a file,
or somewhere else, but I don't think it justifies an entirely separate data
store.