Adopted and enforced code style from other iNat Javascript projects
These rules are largely based on the AirBNB ones, which are not quite standard
for the React Native world, where prettier seems to be more common, but I
think they add a lot of useful checks, and unlike prettier we can customize
them. This also just makes it easier for people on the iNat team to work on
the mobile app.