Fixed messages test; changed useQuery mocking approach
In component unit tests we really don't need to know that useQuery works, or
that our API wrappers work, we just need to know that the component renders
what it's supposed to given data that it receives. So instead of making a
fake hook to set the React Query client cache and waiting for a mocked
network response to complete, I'm just mocking useQuery to return what we
want.

If we do want to test that all these things work together, we should write an
integration test. Setup is more complicated, but that's why you don't write
as many integration tests.