Clear the React Query cache on sign out
Another approach I tried was to add the user ID to the query key in
useAuthenticatedQuery, but we can't synchronously fetch the current user ID
at present, and the hook must be synchronous (I think), so I ended uyp always
performing two queries, one for a signed out state and one for signed in...
every time I used useAuthenticatedQuery. Not good for performance, and
guaranteed to error out for any point that requires auth.