WIP: Realm 11 / React Native 0.70 upgrade
So far I can build for Android and the app doesn't crash, but Realm init fails
in tests with `Class 'Comment' must extend Realm.Object`. Comment *does*
extend Realm.Object so I'm a bit stumped. Using exactly the same model in a
test repo does not raise the error, so I'm wondering if it has something to
do with ES6 transpilation.