Upgrading some dependencies, including realm to 11.0.0
Not sure if all it needed was some time. However, I only upgraded some of the dependencies with minor version number changes, and a few with major for which I knew from Seek that there were no breaking changes to look out for. Most notably I replaced realm with the stable version 11.0.0 and updated react-native-vision-camera. After that both Android and iOS were compiling fine for me.
Have not tested on real devices though.