ObsEdit loads obs from db only when required
This should allow ObsEdit to show the observation specified in the
params *and* not reset its state from the context when returning to ObsEdit
from the Add ID.

I don't love the integration test I added. I wish there was a better way to
assert claims about the state of the context / provider.