Removed custom table styling; removed unnecessary tests
The table rendering was causing a weird segfault on my Pixel 3. We might want
to reintroduce it eventually, but for the sake of stability I think the
default styling is fine for now.

The tests were mainly ones that check for the existence of an element when we
test for the presence of an attribute of that element later in the test. The
latter tests will fail if those elements don't exist.