Several changes to try and address mysterious signouts (#373)
* Configure Realm with the full path to the file
* Remove all Realm.open calls in AuthenticationService in favor of passing the
  context/provider copy of realm (only one realm instance)
* Only delete the realm file on sign out if deleting realm contents fails for
  some reason

This also adds more logging around this issue, and expands the About page to
show the log contents. I also removed some parts of AuthenticationService
that seemed redundant with some of our hooks and that opened separate
connections to Realm, which I thought might be related to the bug in some
way.