Add icons files to repo; move away from Fontastic
This uses an npm package to convert SVGs into an icon font. allowing us to
note step on each others toes when adding new icons in branches and
simplifying the icon addition process. It also allows us to track changes to
the icons, and embed them directly as SVGs if we need to.