Add metro bundler flag
By setting an env var flag MOCK_MODE=e2e the metro bundler accepts files with the extension "xx.e2e-mock.js" over default source files, like "xx.js". This is the recommended way of mocking in the detox documentation. The env var has to be set in the shell in which the actual packager is running, or the build script is running which starts the packager as well.
I have added the setting and unsetting of the env var in the detox e2e build scripts, so for the release mode that should be enough. The only downside is that if someone starts an e2e test in debug mode they have to set the env var in the shell they run the packager alongside the test. This might lead to the env var being set in later builds in the same shell. I have added a check to Fastfile to make sure we at least don't build any apps with the env var.