Create fetchUserLocation.e2e-mock.ts
This is adding the actual mock of location fetching. By overriding the function in a file that will get bundled instead of the actual .js file when the env var is set.
I have added a timeout to the mock to simulate some fetching time for getting the location but this can also be a direct return.