Remove increment from upload event emitters
This attempts to further simplify the upload event emitters by removing the
increment, which never changes, and which should probably be controlled by
the listener, not the emitter, since the listener is responsible for actually
updating the progress. Also tries to ensure that events that get emitted to
signal that an upload finished always happen even if the upload errors out.