WIP: Allow PermissionGate to be used outside of nav hierarchy
This provides a way to show the permission gate in cases where it doesn't need
to block access to an entire screen, like on Explore, where we only need to
show it when the user wants to view their current location.

Not entirely sure I love this approach, but it mostly works.

TODO: figure out how to get useUserLocation to try again after permission has
been granted.