Refactor BottomSheet to always use dynamic snap points
For all usages in the app of the BottomSheet component, or components derived from it. This removes the possibility to specify snapPoints as a prop.
In other words, all instances of bottom sheet will scale, depending on the content to display.
This necessitated some changes in paddings at the bottom of existing bottom sheets.