Made MediaViewer a modal
This takes the MediaViewer out of the navigation hierarchy and makes it a
modal that appears over another navigation screen. React Navigation's modal
presentation doesn't help us here because it doesn't work in Android.

Note that this presents the problem of displaying other modal UI on top of a
fullscreen modal, which is generally impossible in React Native. Here I've
dealt with bottom sheets by making them optionally not display as modals, but
this will also probably present a problem for alerts and such.