Data and layout changes leading up to sound player on ObsDetail
Tries to fetch the sound data and put it in the database, adds sound slide
(sans player) to obs detail. Note there is a significant bug here in which we
are inserting the wrong UUID into the database for the ObservationSound.
We'll need https://github.com/inaturalist/iNaturalistAPI/issues/422 to be
done before we can fix that.