WIP: support multi-upload cancellation
Uses an altered version of inaturalistjs to pass in an AbortController signal
that will be passed through to fetch requests and can be used to cancel those
requests.

TODO: figure out how to support this for the UPLOAD NOW button, which uploads
outside of the MyObs context and thus doesn't have access to its
AbortController.