Rel paths instead of ext so e2e-mock will be found in e2e tests
e2e tests were failing in GitHub actions because the e2e-mock we have for
fetchUserLocation wasn't being used. I think it was because I converted it to
Typescript, and to satisfy eslint I had to add an explicit .ts extension. I
tried a few wasy to address this, but just using a relative path might be the
simplest.