Spread suggestion's taxon into new function return
This is necessary because the representative_photo returned by the image similarity search from API is fleeting only (e.g. from useOnlineSuggestions -> useSuggestions -> suggestions.topSuggestion).
The key is not saved to realm with the taxon because it only is relevant to the current photo scored on the API.
Therefore when overriding the taxon info with offline data we need to keep it.