Disable fetching offline suggestions for successful but empty online results (#2810)
As discussed on the linked linear issue, we want to trust online results more than offline results. So, if the online API call was successful but resulted in zero suggestions (currently, this can happen for example if human is part of the low-confidence suggestions that are filtered out before returning).