Comment out Realm.setLogLevel
I have seen the log category (which is the second, and optional argument here) being referenced in the stacktrace of a Signal 11 crash. I don't know if this call here is related to the crash, but it is worth trying to silence it for a while to see if we get rid of the crash that way.