Consistent location permission asks (#3032)
* Reorder conditions

* This does not really do anything because the option

keys have different names

* The Add Location CTA should always ask for permission

The only state we show the Add Location CTA is if the observation does not have a location when landing on ObsEdit.

* Use an initial state wether an observation should watch users location

In this case we can override the state to true whenever we want to force the app to watch the location.
For example when the user grants location permission on the screen we want to watch no matter where the user came from.

* Remove unused or nowhere imported functions

* Use same navigation on Match as ObsEdit does

* Always ask for permission when the MyLocation button

is pressed on the map

* Add a loading state to the "Add Location button"