Update `LoginSignUpWrapper` to gracefully handle screen size changes.
Prior to this pull request, the `LoginSignUpWrapper` would fetch the screen size _once_, and use that to determine the minimum height of the scroll view content. Because screen sizes are dynamic (e.g. rotation, or resizing apps), this can cause isses with the layout math.

After this pull request, this component will dynamically recalculate the screen size math whenever the screen dimensions change.