Remove ExploreFiltersModal from ExploreRedesigned
Removing this now in preparation of making the Explore Filters a standalone screen called "Advanced Search". However, looking through the code, I see that the filters modal is actually tied to the same ExploreContext as the instance of the screen it is displayed in.
The idea behind making the modal into a standalone screen was to make our lives as developers easier and to not having to have a modal that allows navigating forward (which is unusual for a modal). So, there is a follow-up ticket after this PR to asses how much work it would be to connect two standalone screens to the same context knowing that several instances of those screens can be in the stack simultaneously.